% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_test_data}
\alias{get_test_data}
\title{Get test data that can be used with visualize_data_pairs(data) function}
\usage{
get_test_data()
}
\description{
Get test data that can be used with visualize_data_pairs(data) function
}
\examples{
get_test_data()
}
