% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{visualize_data_pairs}
\alias{visualize_data_pairs}
\title{Visualize data pairs using Shiny application}
\usage{
visualize_data_pairs(data)
}
\arguments{
\item{data}{Data used for visualization. Has to contain columns E1_CONCEPT_ID and E2_CONCEPT_ID}
}
\description{
Visualize data pairs using Shiny application
}
\examples{
visualize_data_pairs(data)
}
